src/federated/client/requirements.txt
flwr
tensorflow
pandas
sklearn
Pillow
Keras-Applications
numpy==1.20.1
matplotlib
psutil
memory-profiler
scipy==1.4.1